Re-ordered site to a more structured format. Re-fractored code.
