+++ package.json
@@ -19,6 +19,7 @@
     "@types/node": "^16.11.36",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
+    "framer-motion": "^6.3.3",
     "graphql": "^16.5.0",
     "node-sass": "^7.0.1",
     "react": "^18.1.0",
