+++ src/components/Filter/Types.ts
@@ -1,0 +1,4 @@
+export enum FilterOptions {
+    OPEN = "OPEN",
+    CLOSED = "CLOSED",
+  }
